chore: modularized application
