+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:orientation_app/app.dart';
 import 'package:orientation_app/constants/routes.dart';
 import 'package:orientation_app/controllers/activities_controller.dart';
 import 'package:orientation_app/controllers/attendance_controller.dart';
@@ -9,8 +10,6 @@ import 'package:orientation_app/controllers/notifications_controller.dart';
 import 'package:orientation_app/controllers/posts_controller.dart';
 import 'package:orientation_app/controllers/statistic_controller.dart';
 import 'package:orientation_app/controllers/usercontroller.dart';
-import 'package:orientation_app/pages/home_page.dart';
-import 'package:orientation_app/pages/splash_screen.dart';
 import 'package:onesignal_flutter/onesignal_flutter.dart';
 
 void main() async {
@@ -29,7 +28,7 @@ void main() async {
 // Enable verbose logging for debugging (remove in production)
     OneSignal.Debug.setLogLevel(OSLogLevel.verbose);
     // Initialize with your OneSignal App ID
-    OneSignal.initialize("38534632-4801-4630-83b4-247fa15d02af");
+    OneSignal.initialize("7fac2b4a-4d01-4665-ad44-ee8497ca8922");
   }
 
   runApp(
@@ -48,18 +47,4 @@ void main() async {
       getPages: AppRoutes.routes,
     ),
   );
-}
-
-class OrientationApp extends StatelessWidget {
-  const OrientationApp({super.key});
-  @override
-  Widget build(BuildContext context) {
-    // Getting user controller
-    UserController userController = Get.find<UserController>();
-    return Obx(
-      () => userController.user.value != null
-          ? const HomePage()
-          : const SplashScreen(),
-    );
-  }
 }
